Add some features and bug fixes for the CampaignSubCategoryController.

+++ src/main/java/com/dxvalley/crowdfunding/controllers/CampaignSubCategoryController.java
@@ -2,6 +2,10 @@ package com.dxvalley.crowdfunding.controllers;
 
 import java.util.List;
 
+import com.dxvalley.crowdfunding.services.CampaignCategoryService;
+import lombok.RequiredArgsConstructor;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -15,41 +19,72 @@ import com.dxvalley.crowdfunding.models.CampaignSubCategory;
 import com.dxvalley.crowdfunding.services.CampaignSubCategoryService;
 
 @RestController
+@RequiredArgsConstructor
 @RequestMapping("/api/campaignSubCategory")
 public class CampaignSubCategoryController {
   private final CampaignSubCategoryService campaignSubCategoryService;
-
-  public CampaignSubCategoryController(CampaignSubCategoryService campaignSubCategoryService) {
-    this.campaignSubCategoryService = campaignSubCategoryService;
-  }
+  private final CampaignCategoryService campaignCategoryService;
 
   @GetMapping("/getSubCategories")
-  List<CampaignSubCategory> gCampaignSubcategories() {
+  List<CampaignSubCategory> getCampaignSubcategories() {
     return this.campaignSubCategoryService.getCampaignSubCategories();
   }
 
   @GetMapping("/{CampaignSubCategoryId}")
-  CampaignSubCategory getCampaignSubCategory(@PathVariable Long campaignSubCategoryId) {
-    return campaignSubCategoryService.getCampaignSubCategoryById(campaignSubCategoryId);
+  ResponseEntity<?> getCampaignSubCategory(@PathVariable Long CampaignSubCategoryId) {
+    var result =  campaignSubCategoryService.getCampaignSubCategoryById(CampaignSubCategoryId);
+
+    if (result == null){
+      return new ResponseEntity<>("There is no subcategory with this ID.", HttpStatus.BAD_REQUEST);
+    }
+    return new ResponseEntity<>(result, HttpStatus.OK);
   }
 
   @PostMapping("/add")
-  CampaignSubCategory addCampaignSubCategory(@RequestBody CampaignSubCategory CampaignSubCategories) {
-    return campaignSubCategoryService.addCampaignSubCategory(CampaignSubCategories);
+  ResponseEntity<?> addCampaignSubCategory(@RequestBody CampaignSubCategory tempCampaignSubCategories) {
+    //TODO: check existence of category before adding subcategory
+    var campaignSubCategories =  campaignSubCategoryService.addCampaignSubCategory(tempCampaignSubCategories);
+      return new ResponseEntity<>(campaignSubCategories, HttpStatus.OK);
   }
 
   @PutMapping("/edit/{campaignSubCategoryId}")
-  CampaignSubCategory ediCampaignSubCategory(@RequestBody CampaignSubCategory tempSubCampaignCategory,
+  ResponseEntity<?> editCampaignSubCategory(@RequestBody CampaignSubCategory tempSubCampaignCategory,
       @PathVariable Long campaignSubCategoryId) {
-    CampaignSubCategory CampaignSubCategory = this.campaignSubCategoryService
+    CampaignSubCategory campaignSubCategory = this.campaignSubCategoryService
         .getCampaignSubCategoryById(campaignSubCategoryId);
-    CampaignSubCategory.setName(tempSubCampaignCategory.getName());
-    CampaignSubCategory.setDescription(tempSubCampaignCategory.getDescription());
-    return campaignSubCategoryService.editCampaignSubCategory(CampaignSubCategory);
-  }
 
+    if (campaignSubCategory == null){
+      return new ResponseEntity<>("There is no subcategory with this ID.", HttpStatus.BAD_REQUEST);
+    }
+
+    campaignSubCategory.setName(
+            tempSubCampaignCategory.getName() != null && tempSubCampaignCategory.getName().length() > 0?
+                    tempSubCampaignCategory.getName() :
+                    campaignSubCategory.getName());
+
+    campaignSubCategory.setDescription(
+            tempSubCampaignCategory.getDescription() != null && tempSubCampaignCategory.getDescription().length() > 0 ?
+                    tempSubCampaignCategory.getDescription() :
+                    campaignSubCategory.getDescription());
+    //TODO: check existence of category before updating subcategory
+    campaignSubCategory.setCategory(
+            tempSubCampaignCategory.getCategory() != null ?
+                    tempSubCampaignCategory.getCategory() :
+                    campaignSubCategory.getCategory());
+
+    var result =  campaignSubCategoryService.editCampaignSubCategory(campaignSubCategory);
+    return new ResponseEntity<>(result, HttpStatus.OK);
+  }
   @DeleteMapping("/delete/{campaignSubCategoryId}")
-  void deleteCampaignCategory(@PathVariable Long campaignSubCategoryId) {
+  ResponseEntity<?> deleteCampaignCategory(@PathVariable Long campaignSubCategoryId) {
+    CampaignSubCategory campaignSubCategory = this.campaignSubCategoryService
+            .getCampaignSubCategoryById(campaignSubCategoryId);
+
+    if (campaignSubCategory == null){
+      return new ResponseEntity<>("There is no subcategory with this ID.", HttpStatus.BAD_REQUEST);
+    }
     campaignSubCategoryService.deleteCampaignSubCategory(campaignSubCategoryId);
+    return new ResponseEntity<>("Deleted Successfully!", HttpStatus.OK);
+
   }
 }
